+++ packages/cindel/lib/cindel.dart
@@ -4,5 +4,6 @@ export 'package:cindel_annotations/cindel_annotations.dart';
 
 export 'src/cindel.dart';
 export 'src/database.dart';
+export 'src/query.dart';
 export 'src/schema.dart';
 export 'src/typed_collection.dart';